Refactor crudkit-id derive macro; Add 128bit number support; Do not distinguish uuid variants

+++ crudkit-id/crates/derive-crudkit-id/tests/12-error-f32.rs
@@ -5,7 +5,7 @@ use derive_crudkit_id::CkId;
 #[derive(CkId)]
 pub struct Foo {
     #[ck_id(id)]
-    id: i32,
+    id: f32,
 }
 
 fn main() {}
